Docs, errors moved
