+++ drink/src/mock/error.rs
@@ -2,6 +2,7 @@ use thiserror::Error;
 
 use crate::Selector;
 
+/// Error type for mocking operations.
 #[derive(Error, Debug)]
 pub enum MockingError {
     #[error("Message not found (unknown selector: {0:?})")]
