+++ drink/src/mock/mocking_api.rs
@@ -5,6 +5,7 @@ use crate::{
     Sandbox, DEFAULT_GAS_LIMIT,
 };
 
+/// Interface for basic mocking operations.
 pub trait MockingApi<R: Runtime> {
     /// Deploy `mock` as a standard contract. Returns the address of the deployed contract.
     fn deploy_mock(&mut self, mock: ContractMock) -> AccountIdFor<R>;
@@ -16,12 +17,16 @@ pub trait MockingApi<R: Runtime> {
 
 impl<R: Runtime> MockingApi<R> for Sandbox<R> {
     fn deploy_mock(&mut self, mock: ContractMock) -> AccountIdFor<R> {
+        // We have to deploy some contract. We use a dummy contract for that. Thanks to that, we
+        // ensure that the pallet will treat our mock just as a regular contract, until we actually
+        // call it.
         let mock_bytes = wat::parse_str(DUMMY_CONTRACT).expect("Dummy contract should be valid");
         let mock_address = self
             .deploy_contract(
                 mock_bytes,
                 0,
                 vec![],
+                // We have to use a different account ID for each contract.
                 vec![self.mock_counter as u8],
                 R::default_actor(),
                 DEFAULT_GAS_LIMIT,
@@ -45,6 +50,9 @@ impl<R: Runtime> MockingApi<R> for Sandbox<R> {
     }
 }
 
+/// A dummy contract that is used to deploy a mock.
+///
+/// Has a single noop constructor and a single panicking message.
 const DUMMY_CONTRACT: &str = r#"
 (module
 	(import "env" "memory" (memory 1 1))
