+++ drink/src/session/error.rs
@@ -37,29 +37,3 @@ pub enum SessionError {
     #[error("Missing transcoder")]
     NoTranscoder,
 }
-
-/// Every contract message wraps its return value in `Result<T, LangResult>`. This is the error
-/// type.
-///
-/// Copied from ink primitives.
-#[non_exhaustive]
-#[repr(u32)]
-#[derive(
-    Debug,
-    Copy,
-    Clone,
-    PartialEq,
-    Eq,
-    parity_scale_codec::Encode,
-    parity_scale_codec::Decode,
-    scale_info::TypeInfo,
-    Error,
-)]
-pub enum LangError {
-    /// Failed to read execution input for the dispatchable.
-    #[error("Failed to read execution input for the dispatchable.")]
-    CouldNotReadInput = 1u32,
-}
-
-/// The `Result` type for ink! messages.
-pub type MessageResult<T> = Result<T, LangError>;
